2nd version of localFixing and new features
2nd version of localFixing and new features function that featurize's our data.

+++ localFixing.py
@@ -30,6 +30,9 @@ else:
     #change to the correct directory
     os.chdir(os.path.dirname(os.path.realpath(__file__)) +'\PDBFiles')
 
+#get the pdb files directory for later use
+pdbDirec = os.getcwd()
+
 #Now Check for the dataset being downloaded
 if not os.path.exists(os.path.dirname(os.path.realpath(__file__)) +'\pdbbind_v2019_other_PL.tar.gz'):
   wholeSet = input('Would you Like to download whole pdbbind Dataset? \n(Y/N)\n')
@@ -71,13 +74,21 @@ os.chdir(fileDict[dataFile])
 #save the current working directory before i mess with it!
 direc = os.getcwd()
 for pdbid in os.listdir(direc):
-  try:
-    fixer = PDBFixer(url='https://files.rcsb.org/download/%s.pdb' % (pdbid))
-    PDBFile.writeFile(fixer.topology, fixer.positions, open('%s.pdb' % (pdbid), 'w'))
+  #print(pdbid)
+  if not direc == os.getcwd():
+    os.chdir(direc)
+  else:
+    pass
 
+  p, m = None, None
+  try:
+    if not os.path.exists(os.path.dirname(os.path.realpath(__file__)) + '\%s' % (pdbid)):
+      fixer = PDBFixer(url='https://files.rcsb.org/download/%s.pdb' % (pdbid))
+      PDBFile.writeFile(fixer.topology, fixer.positions, open('%s.pdb' % (pdbid), 'w'))
     #write all of the ligand files as well
     #WE WANT SMILES STUFF BEFORE  WE SAVE THE FILE
     os.chdir(direc+'/%s' % (pdbid))
+    ligand = None
     for eachfile in os.listdir(os.getcwd()):
       if eachfile == ('%s_ligand.mol2' % (pdbid)):
         molfile = Chem.rdmolfiles.MolFromMol2File(eachfile)
@@ -89,15 +100,40 @@ for pdbid in os.listdir(direc):
     #return to the previous directory
     os.chdir(direc)
 
+    if not os.path.exists(os.path.dirname(os.path.realpath(__file__)) + '\Processedfiles.txt'):
+      newfile = open('Processedfiles.txt', 'w')
+      newfile.close()
+
     #now run the full fixing step
-    p, m = None
     try:
-      p, m = prepare_inputs('%s.pdb' % (pdbid), ligand)
+      newFile = open('Processedfiles.txt', 'r')
+      if str(pdbid) not in newFile.read():
+        newFile.close()
+        print("Sanitizing %s" %(pdbid))
+        proFile = open('Processedfiles.txt', 'a')
+        proFile.write(str(pdbid+'\n'))
+        proFile.close()
+        p, m = prepare_inputs('%s.pdb' % (pdbid), ligand)
+        newFile.close()
+      else:
+        continue
     except:
       print('%s failed sanitization' % (pdbid))
   except:
     continue
-  
+    
   if p and m:  # protein and molecule are readable by RDKit
-    Chem.rdmolfiles.MolToPDBFile(p, '%s.pdb' % (pdbid))
-    Chem.rdmolfiles.MolToPDBFile(m, 'ligand_%s.pdb' % (pdbid))
+    #Start in the directory of pdb files
+    os.chdir(pdbDirec)
+
+    #make a new file called 'sanitized files"and switch to it
+    if not os.path.exists(os.path.dirname(os.path.realpath(__file__)) +'\SanitizedFiles'):
+      os.makedirs(os.path.dirname(os.path.realpath(__file__)) +'\SanitizedFiles')
+    else:
+      pass
+    os.chdir(os.getcwd()+'\SanitizedFiles')
+
+    #check if there is already a sanitized file present. if not, make one. if so, skip it
+    if not (os.path.exists(os.path.dirname(os.path.realpath(__file__)) + '\%s.pdb' % (pdbid)) and os.path.exists(os.path.dirname(os.path.realpath(__file__)) + '\ligand_%s.pdb' % (pdbid))):
+      Chem.rdmolfiles.MolToPDBFile(p, '%s.pdb' % (pdbid))
+      Chem.rdmolfiles.MolToPDBFile(m, 'ligand_%s.pdb' % (pdbid))
